Moved propagate service to domain from infra.
