+++ consensus/pbft/propagate_service.go
@@ -14,14 +14,13 @@
  * limitations under the License.
  */
 
-package adapter
+package pbft
 
 import (
 	"errors"
 
 	"github.com/it-chain/engine/common"
 	"github.com/it-chain/engine/common/command"
-	"github.com/it-chain/engine/consensus/pbft"
 	"github.com/rs/xid"
 )
 
@@ -30,19 +29,17 @@ var ErrEmptyBlock = errors.New("Block is empty")
 var ErrEmptyBlockHash = errors.New("Block hash is empty")
 var ErrEmptyMsg = errors.New("Message is empty")
 
-type Publish func(topic string, data interface{}) (err error)
-
 type PropagateService struct {
-	publish Publish
+	eventService common.EventService
 }
 
-func NewPropagateService(publish Publish) PropagateService {
+func NewPropagateService(eventService common.EventService) PropagateService {
 	return PropagateService{
-		publish: publish,
+		eventService: eventService,
 	}
 }
 
-func (ps PropagateService) BroadcastProposeMsg(msg pbft.ProposeMsg, representatives []*pbft.Representative) error {
+func (ps PropagateService) BroadcastProposeMsg(msg ProposeMsg, representatives []*Representative) error {
 	if msg.StateID.ID == "" {
 		return ErrStateIdEmpty
 	}
@@ -64,7 +61,7 @@ func (ps PropagateService) BroadcastProposeMsg(msg pbft.ProposeMsg, representati
 	return nil
 }
 
-func (ps PropagateService) BroadcastPrevoteMsg(msg pbft.PrevoteMsg, representatives []*pbft.Representative) error {
+func (ps PropagateService) BroadcastPrevoteMsg(msg PrevoteMsg, representatives []*Representative) error {
 	if msg.StateID.ID == "" {
 		return ErrStateIdEmpty
 	}
@@ -86,7 +83,7 @@ func (ps PropagateService) BroadcastPrevoteMsg(msg pbft.PrevoteMsg, representati
 	return nil
 }
 
-func (ps PropagateService) BroadcastPreCommitMsg(msg pbft.PreCommitMsg, representatives []*pbft.Representative) error {
+func (ps PropagateService) BroadcastPreCommitMsg(msg PreCommitMsg, representatives []*Representative) error {
 	if msg.StateID.ID == "" {
 		return ErrStateIdEmpty
 	}
@@ -104,22 +101,22 @@ func (ps PropagateService) BroadcastPreCommitMsg(msg pbft.PreCommitMsg, represen
 	return nil
 }
 
-func (ps PropagateService) broadcastMsg(SerializedMsg []byte, protocol string, representatives []*pbft.Representative) error {
+func (ps PropagateService) broadcastMsg(SerializedMsg []byte, protocol string, representatives []*Representative) error {
 	if SerializedMsg == nil {
 		return ErrEmptyMsg
 	}
 
-	command, err := createDeliverGrpcCommand(protocol, SerializedMsg)
+	grpcCommand, err := createDeliverGrpcCommand(protocol, SerializedMsg)
 
 	if err != nil {
 		return err
 	}
 
 	for _, r := range representatives {
-		command.RecipientList = append(command.RecipientList, r.GetID())
+		grpcCommand.RecipientList = append(grpcCommand.RecipientList, r.GetID())
 	}
 
-	return ps.publish("message.deliver", command)
+	return ps.eventService.Publish("message.deliver", grpcCommand)
 }
 
 func createDeliverGrpcCommand(protocol string, body interface{}) (command.DeliverGrpc, error) {
